don't make bin required, sometimes it's already in the bin directory

+++ zip-it.sh
@@ -4,8 +4,8 @@ sed -i "s/%package%/${INPUT_PACKAGE_NAME}/g" /zip-it;
 sed -i "s|%prefix%|${INPUT_PREFIX}|g" /crackle.conf;
 sed -i "s/%package%/${INPUT_PACKAGE_NAME}/g" /crackle.conf;
 sed -i "s/%maintainer%/${INPUT_MAINTAINER}/g" /crackle.conf;
-mv /zip-it ${INPUT_PACKAGE_NAME}|| exit 1;
-mv $INPUT_BIN bin/
+mv /zip-it ${INPUT_PACKAGE_NAME}
+[[ -e $INPUT_BIN ]] && mv $INPUT_BIN bin/
 [[ -e $INPUT_LIB ]] && mv $INPUT_LIB lib/
 [[ -e $INPUT_SHARE ]] && mv $INPUT_SHARE share
 zip -r $INPUT_PACKAGE_NAME.zip $INPUT_PKG_DIRECTORIES ${INPUT_PACKAGE_NAME} /crackle.conf
